Implement backend event cancellation machinery
This is a breaking API change. Rather than using the 'date' field
in the passed JSON object for setting and removing EventTimes, this
change uses a new field called 'dateStatuses', which is an array of
dateStatus objects of the following format:

{
    'id': <EventTime pkid || null>,
    'date': <EventTime date>,
    'status': < 'C' || 'A' >,
    'newsflash': <Newsflash shown if status === 'C' || null>
}

The 'id' field should be null for a new EventTime, or the pkid for
existing EventTimes.

The 'date' field should be of the same format as the old date field:
'yyyy-mm-dd'.

The status should be 'A' for normal or 'C' for cancelled. It may be left
null for new events, in which case it will default to 'A'.

The newsflash should be populated with a message if status === 'C',
otherwise it may be populated with either a message or null.

To delete an existing EventTime, simply remove it's corresponding
dateStatus object from the dateStatuses field.

Also in this commit: refactoring manage_event.php to make it more
readable.